Added different colors to teams
